requirements.txt
requests>=2.28.0
python-dotenv>=0.21.0
tenacity>=8.2.0